chore: generalize subtask.
